Fix Next.js build: Suspense around useSearchParams, lazy Resend init; include .env by removing ignore; add client wrappers
