+++ src/app/dashboard/access/page.tsx
@@ -1,7 +1,5 @@
-"use client";
-
-import { useEffect } from "react";
-import { useRouter, useSearchParams } from "next/navigation";
+import { Suspense } from "react";
+import AccessContent from "./AccessContent";
 import {
   Card,
   CardContent,
@@ -9,77 +7,35 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
-import { Shield, Loader2, AlertCircle } from "lucide-react";
-import { Button } from "@/components/ui/button";
-import Link from "next/link";
+import { Loader2 } from "lucide-react";
 
-export default function DashboardAccessPage() {
-  const router = useRouter();
-  const searchParams = useSearchParams();
+export const dynamic = "force-dynamic";
 
-  useEffect(() => {
-    const token = searchParams.get("token");
-
-    if (token) {
-      // Redirect to dashboard with token
-      router.replace(`/dashboard?token=${token}`);
-    } else {
-      // If no token, redirect to waitlist after a delay
-      setTimeout(() => {
-        router.replace("/waitlist");
-      }, 3000);
-    }
-  }, [searchParams, router]);
-
-  const token = searchParams.get("token");
-
-  if (token) {
-    return (
-      <div className="min-h-screen bg-background flex items-center justify-center">
-        <Card className="w-[400px]">
-          <CardHeader className="text-center">
-            <div className="flex items-center justify-center space-x-2 mb-4">
-              <Shield className="h-8 w-8 text-primary" />
-              <span className="text-2xl font-bold">PayPerCrawl</span>
-            </div>
-            <CardTitle>Redirecting to Dashboard</CardTitle>
-            <CardDescription>
-              Please wait while we redirect you to your beta dashboard...
-            </CardDescription>
-          </CardHeader>
-          <CardContent className="flex justify-center">
-            <Loader2 className="h-8 w-8 animate-spin text-primary" />
-          </CardContent>
-        </Card>
-      </div>
-    );
-  }
-
+function AccessFallback() {
   return (
     <div className="min-h-screen bg-background flex items-center justify-center">
       <Card className="w-[400px]">
         <CardHeader className="text-center">
           <div className="flex items-center justify-center space-x-2 mb-4">
-            <AlertCircle className="h-8 w-8 text-destructive" />
             <span className="text-2xl font-bold">PayPerCrawl</span>
           </div>
-          <CardTitle>Access Denied</CardTitle>
+          <CardTitle>Processing Access</CardTitle>
           <CardDescription>
-            No valid invitation token found. You need an invitation to access
-            the dashboard.
+            Please wait while we process your invitation token...
           </CardDescription>
         </CardHeader>
-        <CardContent className="space-y-4">
-          <p className="text-sm text-muted-foreground text-center">
-            Redirecting to waitlist in a few seconds...
-          </p>
-          <div className="flex justify-center">
-            <Link href="/waitlist">
-              <Button>Join Waitlist</Button>
-            </Link>
-          </div>
+        <CardContent className="flex justify-center">
+          <Loader2 className="h-8 w-8 animate-spin text-primary" />
         </CardContent>
       </Card>
     </div>
+  );
+}
+
+export default function DashboardAccessPage() {
+  return (
+    <Suspense fallback={<AccessFallback />}>
+      <AccessContent />
+    </Suspense>
   );
 }
